Unified and enforced test conventions.
