+++ .circleci/template.yml
@@ -94,6 +94,9 @@ jobs:
           name: Validate versions of dependencies
           command: pnpm run check-versions-match
       - run:
+          name: Validate naming and location of test files
+          command: pnpm run check-test-conventions
+      - run:
           name: Type check the project
           command: pnpm run typecheck
       - run:
