+++ package.json
@@ -19,6 +19,7 @@
     "@inquirer/prompts": "^7.10.1",
     "@listr2/prompt-adapter-inquirer": "^2.0.22",
     "@vitest/coverage-v8": "^4.1.2",
+    "@vitest/eslint-plugin": "^1.6.24",
     "eslint": "^10.0.0",
     "eslint-config-ckeditor5": "^19.0.0",
     "eslint-plugin-ckeditor5-rules": "^19.0.0",
@@ -43,6 +44,7 @@
     "test:scripts": "vitest --config ./scripts-tests/vitest.config.mjs --run",
     "coverage": "node ./scripts/runtest.js --coverage",
     "check-versions-match": "node scripts/ci/check-dependencies-versions-match.mjs",
+    "check-test-conventions": "node scripts/ci/check-test-file-conventions.mjs",
     "release:prepare-changelog": "node ./scripts/preparechangelog.js",
     "release:prepare-packages": "node ./scripts/preparepackages.js",
     "release:publish-packages": "node ./scripts/publishpackages.js",
