+++ packages/ckeditor5-dev-build-tools/tests/plugins/banner/banner.ts
@@ -1,0 +1,165 @@
+/**
+ * @license Copyright (c) 2003-2026, CKSource Holding sp. z o.o. All rights reserved.
+ * For licensing, see LICENSE.md.
+ */
+
+import { join } from 'node:path';
+import { describe, expect, it, vi } from 'vitest';
+import { rolldown, type RolldownOutput, type OutputAsset, type Plugin } from 'rolldown';
+import { verifyAsset, verifyChunk } from '../../_utils/utils.js';
+import { addBanner, bundleCss, type RollupBannerOptions } from '../../../src/index.js';
+
+const createFilterSpy = vi.hoisted( vi.fn );
+
+vi.mock( '@rollup/pluginutils', async importOriginal => {
+	const original = await importOriginal() as any;
+
+	return {
+		...original,
+		createFilter: ( ...args: any ) => {
+			createFilterSpy( ...args );
+			return original.createFilter( ...args );
+		}
+	};
+} );
+
+/**
+ * Helper function for creating a bundle that won't be written to the file system.
+ */
+async function generateBundle( options: RollupBannerOptions, sourcemap: boolean = false ): Promise<RolldownOutput[ 'output' ]> {
+	const bundle = await rolldown( {
+		input: join( import.meta.dirname, './fixtures/input.ts' ),
+		plugins: [
+			bundleCss( {
+				fileName: 'styles.css',
+				sourceMap: sourcemap
+			} ),
+
+			addBanner( options )
+		]
+	} );
+
+	const { output } = await bundle.generate( {
+		format: 'esm',
+		file: 'input.js',
+		assetFileNames: '[name][extname]',
+		sourcemap
+	} );
+
+	return output;
+}
+
+function emitMapWithoutFileProperty(): Plugin {
+	return {
+		name: 'emit-map-without-file-property',
+
+		generateBundle() {
+			this.emitFile( {
+				type: 'asset',
+				fileName: 'custom.css',
+				source: '.x { color: red; }'
+			} );
+
+			this.emitFile( {
+				type: 'asset',
+				fileName: 'custom.css.map',
+				source: JSON.stringify( {
+					version: 3,
+					sources: [ 'custom.css' ],
+					sourcesContent: [ '.x { color: red; }' ],
+					names: [],
+					mappings: 'AAAA'
+				} )
+			} );
+		}
+	};
+}
+
+describe( 'addBanner()', () => {
+	it( 'Adds banner to .js and .css files by default', async () => {
+		const banner = '/* CUSTOM BANNER */';
+		const output = await generateBundle( { banner } );
+
+		verifyChunk( output, 'input.js', banner );
+		verifyAsset( output, 'styles.css', banner );
+		verifyAsset( output, 'styles-editor.css', banner );
+		verifyAsset( output, 'styles-content.css', banner );
+	} );
+
+	it( 'Updates the source map', async () => {
+		const unmodifiedOutput = await generateBundle( { banner: '' }, true );
+		const banner = '/* CUSTOM BANNER */\n';
+		const output = await generateBundle( { banner }, true );
+
+		const unmodifiedSourceMap = unmodifiedOutput.find( item => item.fileName === 'styles-editor.css.map' ) as OutputAsset;
+		const sourceMap = output.find( item => item.fileName === 'styles-editor.css.map' ) as OutputAsset;
+
+		expect( unmodifiedSourceMap.source ).not.toBe( sourceMap.source );
+
+		verifyChunk( output, 'input.js', banner );
+		verifyAsset( output, 'styles.css', banner );
+		verifyAsset( output, 'styles-editor.css', banner );
+		verifyAsset( output, 'styles-content.css', banner );
+	} );
+
+	it( 'Allows overriding "include" option', async () => {
+		const banner = '/* CUSTOM BANNER */';
+		const output = await generateBundle( {
+			banner,
+			include: [ '**/*.css' ]
+		} );
+
+		expect( output[ 0 ].code ).not.includes( banner );
+		verifyAsset( output, 'styles.css', banner );
+	} );
+
+	it( 'Allows overriding "exclude" option', async () => {
+		const banner = '/* CUSTOM BANNER */';
+		const output = await generateBundle( {
+			banner,
+			exclude: [ '**/*.js' ]
+		} );
+
+		expect( output[ 0 ].code ).not.includes( banner );
+		verifyAsset( output, 'styles.css', banner );
+	} );
+
+	it( 'Should have proper default values', async () => {
+		const banner = '/* CUSTOM BANNER */';
+		await generateBundle( { banner } );
+
+		expect( createFilterSpy ).toHaveBeenCalledExactlyOnceWith(
+			[
+				'**/*.js',
+				'**/*.css',
+				'**/translations/**/*.d.ts'
+			],
+			null
+		);
+	} );
+
+	it( 'Handles source maps without the "file" property', async () => {
+		const banner = '/* CUSTOM BANNER */\n';
+
+		const bundle = await rolldown( {
+			input: join( import.meta.dirname, './fixtures/input.ts' ),
+			plugins: [
+				emitMapWithoutFileProperty(),
+				bundleCss( {
+					fileName: 'styles.css'
+				} ),
+				addBanner( { banner } )
+			]
+		} );
+
+		const { output } = await bundle.generate( {
+			format: 'esm',
+			file: 'input.js',
+			assetFileNames: '[name][extname]',
+			sourcemap: true
+		} );
+
+		verifyAsset( output, 'custom.css', banner );
+		verifyAsset( output, 'custom.css.map', '"version":3' );
+	} );
+} );
