+++ packages/ckeditor5-dev-build-tools/tests/plugins/loadSourcemaps/loadSourcemaps.ts
@@ -4,7 +4,7 @@
  */
 
 import { join } from 'node:path';
-import { test, expect } from 'vitest';
+import { describe, expect, it } from 'vitest';
 import { rolldown, type RolldownOutput, type OutputAsset } from 'rolldown';
 
 import { loadSourcemaps } from '../../../src/index.js';
@@ -28,13 +28,15 @@ async function generateBundle( input: string, sourcemap: boolean = false ): Prom
 	return output;
 }
 
-test( 'Emits source maps combined with source maps of dependencies', async () => {
-	const output = await generateBundle( './fixtures/input.ts', true );
-	const sourceMap = output.find( asset => asset.fileName === 'input.js.map' ) as OutputAsset;
+describe( 'loadSourcemaps()', () => {
+	it( 'Emits source maps combined with source maps of dependencies', async () => {
+		const output = await generateBundle( './fixtures/input.ts', true );
+		const sourceMap = output.find( asset => asset.fileName === 'input.js.map' ) as OutputAsset;
 
-	/**
-	 * The resulting source map will only contain the `dependency.ts` string if the source map
-	 * of the `dependency.js` fixture was loaded and combined with the source map of the input file.
-	 */
-	expect( sourceMap.source ).toContain( 'dependency.ts' );
+		/**
+		 * The resulting source map will only contain the `dependency.ts` string if the source map
+		 * of the `dependency.js` fixture was loaded and combined with the source map of the input file.
+		 */
+		expect( sourceMap.source ).toContain( 'dependency.ts' );
+	} );
 } );
