+++ packages/ckeditor5-dev-build-tools/vitest.config.ts
@@ -13,6 +13,13 @@ export default defineConfig( {
 		mockReset: true,
 		unstubEnvs: true,
 		unstubGlobals: true,
+		include: [
+			'tests/**/*.ts'
+		],
+		exclude: [
+			'tests/_utils/**/*.ts',
+			'tests/**/fixtures/**/*.ts'
+		],
 		coverage: {
 			provider: 'v8',
 			include: [
