+++ packages/ckeditor5-dev-manual-server/tests/theme/catalog-search.ts
@@ -3,7 +3,7 @@
  * For licensing, see LICENSE.md.
  */
 
-import { describe, expect, test } from 'vitest';
+import { describe, expect, it } from 'vitest';
 import { filterEntries, findMatchOffsets } from '../../theme/catalog-search.js';
 
 describe( 'filterEntries()', () => {
@@ -25,30 +25,30 @@ describe( 'filterEntries()', () => {
 		}
 	];
 
-	test( 'returns all entries for an empty query', () => {
+	it( 'returns all entries for an empty query', () => {
 		expect( filterEntries( entries, '' ) ).to.equal( entries );
 	} );
 
-	test( 'returns all tests belonging to a matching package', () => {
+	it( 'returns all tests belonging to a matching package', () => {
 		expect( filterEntries( entries, 'foo' ) ).to.deep.equal( entries.slice( 0, 2 ) );
 	} );
 
-	test( 'returns tests with a matching slug, including nested slugs', () => {
+	it( 'returns tests with a matching slug, including nested slugs', () => {
 		expect( filterEntries( entries, 'advanced' ) ).to.deep.equal( [ entries[ 1 ] ] );
 	} );
 
-	test( 'returns no entries when the query does not match', () => {
+	it( 'returns no entries when the query does not match', () => {
 		expect( filterEntries( entries, 'missing' ) ).to.deep.equal( [] );
 	} );
 
-	test( 'preserves original entry objects and their additional fields', () => {
+	it( 'preserves original entry objects and their additional fields', () => {
 		const [ result ] = filterEntries( entries, 'advanced' );
 
 		expect( result ).to.equal( entries[ 1 ] );
 		expect( result!.href ).to.equal( '/packages/ckeditor5-foo/tests/manual/nested/advanced.manual.html' );
 	} );
 
-	test( 'ranks exact, prefix and partial package matches before slug-only matches', () => {
+	it( 'ranks exact, prefix and partial package matches before slug-only matches', () => {
 		const entries = [
 			{ packageName: 'ckeditor5-bar', slug: 'foo-test' },
 			{ packageName: 'ckeditor5-extra-foo', slug: 'unrelated' },
@@ -59,7 +59,7 @@ describe( 'filterEntries()', () => {
 		expect( filterEntries( entries, 'foo' ) ).to.deep.equal( [ entries[ 3 ], entries[ 2 ], entries[ 1 ], entries[ 0 ] ] );
 	} );
 
-	test( 'ranks slug-only package matches by the number of matching tests', () => {
+	it( 'ranks slug-only package matches by the number of matching tests', () => {
 		const entries = [
 			{ packageName: 'ckeditor5-bar', slug: 'foo-test' },
 			{ packageName: 'ckeditor5-baz', slug: 'foo-basic' },
@@ -69,7 +69,7 @@ describe( 'filterEntries()', () => {
 		expect( filterEntries( entries, 'foo' ) ).to.deep.equal( [ entries[ 1 ], entries[ 2 ], entries[ 0 ] ] );
 	} );
 
-	test( 'ranks exact package names without the ckeditor5 prefix', () => {
+	it( 'ranks exact package names without the ckeditor5 prefix', () => {
 		const entries = [
 			{ packageName: 'bar', slug: 'foo-test' },
 			{ packageName: 'foo', slug: 'basic' }
@@ -78,7 +78,7 @@ describe( 'filterEntries()', () => {
 		expect( filterEntries( entries, 'foo' ) ).to.deep.equal( [ entries[ 1 ], entries[ 0 ] ] );
 	} );
 
-	test( 'preserves package and test order when ranking signals are equal', () => {
+	it( 'preserves package and test order when ranking signals are equal', () => {
 		const entries = [
 			{ packageName: 'ckeditor5-foo', slug: 'nested/zeta' },
 			{ packageName: 'ckeditor5-foo', slug: 'nested/alpha' },
@@ -91,7 +91,7 @@ describe( 'filterEntries()', () => {
 } );
 
 describe( 'findMatchOffsets()', () => {
-	test( 'finds matches at the beginning, middle and end of text', () => {
+	it( 'finds matches at the beginning, middle and end of text', () => {
 		expect( findMatchOffsets( 'foo-prefix-foo-suffix-foo', 'foo' ) ).to.deep.equal( [
 			{ start: 0, end: 3 },
 			{ start: 11, end: 14 },
@@ -99,33 +99,33 @@ describe( 'findMatchOffsets()', () => {
 		] );
 	} );
 
-	test( 'finds matches case-insensitively while preserving original offsets', () => {
+	it( 'finds matches case-insensitively while preserving original offsets', () => {
 		expect( findMatchOffsets( 'CKEditor5-ckeditor5', 'ckeditor5' ) ).to.deep.equal( [
 			{ start: 0, end: 9 },
 			{ start: 10, end: 19 }
 		] );
 	} );
 
-	test( 'treats regular expression characters literally', () => {
+	it( 'treats regular expression characters literally', () => {
 		expect( findMatchOffsets( 'foo.[bar]-foo.[bar]', '.[bar]' ) ).to.deep.equal( [
 			{ start: 3, end: 9 },
 			{ start: 13, end: 19 }
 		] );
 	} );
 
-	test( 'returns no matches for an empty or absent query', () => {
+	it( 'returns no matches for an empty or absent query', () => {
 		expect( findMatchOffsets( 'foo', '' ) ).to.deep.equal( [] );
 		expect( findMatchOffsets( 'foo', 'bar' ) ).to.deep.equal( [] );
 	} );
 
-	test( 'finds adjacent matches', () => {
+	it( 'finds adjacent matches', () => {
 		expect( findMatchOffsets( 'foofoo', 'foo' ) ).to.deep.equal( [
 			{ start: 0, end: 3 },
 			{ start: 3, end: 6 }
 		] );
 	} );
 
-	test( 'returns non-overlapping matches', () => {
+	it( 'returns non-overlapping matches', () => {
 		expect( findMatchOffsets( 'aaaa', 'aa' ) ).to.deep.equal( [
 			{ start: 0, end: 2 },
 			{ start: 2, end: 4 }
