+++ packages/ckeditor5-dev-manual-server/vitest.config.ts
@@ -13,6 +13,12 @@ export default defineConfig( {
 		mockReset: true,
 		unstubEnvs: true,
 		unstubGlobals: true,
+		include: [
+			'tests/**/*.ts'
+		],
+		exclude: [
+			'tests/_utils/**/*.ts'
+		],
 		coverage: {
 			provider: 'v8',
 			include: [
