+++ packages/ckeditor5-dev-web-crawler/tests/page/request-policy.ts
@@ -3,7 +3,7 @@
  * For licensing, see LICENSE.md.
  */
 
-import { describe, expect, test } from 'vitest';
+import { describe, expect, it } from 'vitest';
 import type { HTTPRequest } from 'puppeteer';
 import { shouldAbortRequest } from '../../src/page/request-policy.js';
 
@@ -17,31 +17,31 @@ function createRequest( url: string, resourceType: ResourceType = 'document' ): 
 }
 
 describe( 'shouldAbortRequest()', () => {
-	test( 'returns true for media requests', () => {
+	it( 'returns true for media requests', () => {
 		const request = createRequest( 'https://ckeditor.com/video.mp4', 'media' );
 
 		expect( shouldAbortRequest( request ) ).toBe( true );
 	} );
 
-	test( 'returns true for ignored hosts', () => {
+	it( 'returns true for ignored hosts', () => {
 		const request = createRequest( 'https://player.vimeo.com/video/123' );
 
 		expect( shouldAbortRequest( request ) ).toBe( true );
 	} );
 
-	test( 'returns true for api.json files', () => {
+	it( 'returns true for api.json files', () => {
 		const request = createRequest( 'https://ckeditor.com/docs/api.json' );
 
 		expect( shouldAbortRequest( request ) ).toBe( true );
 	} );
 
-	test( 'does not block requests without hostnames', () => {
+	it( 'does not block requests without hostnames', () => {
 		const request = createRequest( 'data:text/plain,hello', 'media' );
 
 		expect( shouldAbortRequest( request ) ).toBe( false );
 	} );
 
-	test( 'returns false for regular requests', () => {
+	it( 'returns false for regular requests', () => {
 		const request = createRequest( 'https://ckeditor.com/docs/guide' );
 
 		expect( shouldAbortRequest( request ) ).toBe( false );
